Add manifest creation and validation commands
The manifest is a JWT token signed with an RSA key containing a `sl` claims for each combination of email/domain and sponsorable account currently being sponsored by the authenticated user.

The list of hashes (claims) are not signed when created locally (since users won't have the private key). During the LinkCommand run, the unsigned manifest would be POSTed to a backend endpoint (which at this point can no longer know what's the source email/domain/sponsorable that generated them) and attempts to validate it.

The only way the validated and signed manifest can be returned is if the user has previously installed the GitHub SponsorLink app (on the web), which asks for explicit email sharing permission. At that point, the server has access to the emails which can be used to validate incoming manifest requests.

If the client-sent manifest contains all entries that can be validated on the backend, the same manifest is returned, but this time properly signed with the SponsorLink private key.

This file lives in the user's personal documents path, and would be the only thing used to check sponsorships.

Since the manifest contains org domains and sponsored accounts, alongside user emails, during the build, the check is a quick Exists(hash(email, sponsorable)) lookup (which can be further optimized via inputs/outputs so it's incremental.

The tests excercise the end to end scenarios using locally-generated (public and private) keys to showcase how it works.

The SponsorLink public key is embedded in the commands library and would likewise be embedded in the analyzer assembly.

Fixes https://github.com/devlooped/SponsorLink/issues/31