Further simplify embeddability of SL single-file
By wrapping all logic and types in a single static class SponsorLink, it's easier to invoke it from multiple places with one-liners.

The main checks would be:
* SponsorLink.Status: to check whether the user opted-in (!NotFound) and refresh is needed (Expired | Invalid)
* SponsorLink.IsSponsoring(user, sponsorable): if status is valid, check whether a user is sponsoring. Null otherwise.

So a quick check to "do nothing unless opted-in" would be to check for the status, or IsSponsoring and check for a null return value.