Add convenience properties to check for IDE
SponsorLink can typically be checked within an IDE to provide enhanced editing experience to sponsors. As such, checking whether certain logic is running in an editor is important.

For now, we provide IsEditor by detecting IsVisualStudio and IsRider. Other editor might be added in the future if we can find a way to detect them.